backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
tensorflow==2.16.1
pillow==10.4.0
numpy==1.26.4
pydantic==2.8.2
